This one was funny, if it's right

+++ src/python_krebsutils/samplevessels.cpp
@@ -605,7 +605,7 @@ double compute_vessel_boxcounts(np::ndarray pypos, np::ndarray pyedges, np::ndar
     Float3 wp[2];
     for (int dim=0; dim<3; ++dim)
       for (int i=0; i<2; ++i)
-        wp[i][dim] = py::extract<float>(pypos[py::extract<int>(pyedges[vi][i])][ dim]);
+        wp[i][dim] = py::extract<float>(pypos[(int) (py::extract<int>(pyedges[vi][i]))][ dim]);
     
     FloatBBox3 vwbb = FloatBBox3().Add(wp[0]).Add(wp[1]);
     vwbb.Extend(py::extract<float>(pyradius[vi]));
